dataframe 0.1
Created new .csv for financial data that python is able to interpret more easily.

Started a dataframe off of the kaggle workbook: "https://www.kaggle.com/samratp/bike-share-analysis/notebook"

This is a dictionary to be imported for future use and  primary purpose is to concatinate/join our data as well as applying financials.